fix(STAR-1514): store and use output name when starting productions
Fix the bug by storing the name of the output that is configured in
the production object. Clean up not found and unused outputs as well.
This avoids errors when a pipeline has been changed to one where some
outputs doesn't exists. This would cause errors that the users would
have a really hard time to fix without altering database or pipeline
configurations.