fix: remove the check for outputs on same ip and port
For SRT callers it can be valid to do multiple streams to same
remote ip and port, the receiver may differentiate the streams
using the stream IDs. For listerners on same machine this type
of check could been ok to have, but not for pipelines on
different machines. Removing the check completely for now.